+++ goflow/engine.go
@@ -1,18 +1,63 @@
 package goflow
 
 import (
+	"crypto/tls"
+	"net/http"
 	"sync"
+	"time"
 
+	"github.com/nyaruka/gocommon/urns"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/flows/engine"
+	"github.com/nyaruka/goflow/services/webhooks"
 	"github.com/nyaruka/mailroom/config"
+
+	"github.com/shopspring/decimal"
 )
 
-// Engine returns the global engine for use in mailroom
+var httpClient *http.Client
+var eng, simulator flows.Engine
+var engInit, simulatorInit sync.Once
+var classificationFactory engine.ClassificationServiceFactory
+var airtimeFactory engine.AirtimeServiceFactory
+
+func init() {
+	// customize the default golang transport
+	t := http.DefaultTransport.(*http.Transport).Clone()
+	t.MaxIdleConns = 32
+	t.MaxIdleConnsPerHost = 8
+	t.IdleConnTimeout = 30 * time.Second
+	t.TLSClientConfig = &tls.Config{
+		Renegotiation: tls.RenegotiateOnceAsClient, // support single TLS renegotiation
+	}
+
+	httpClient = &http.Client{Transport: t, Timeout: time.Duration(15 * time.Second)}
+}
+
+// RegisterClassificationServiceFactory can be used by outside callers to register a classification factory
+// for use by the engine
+func RegisterClassificationServiceFactory(factory engine.ClassificationServiceFactory) {
+	classificationFactory = factory
+}
+
+// RegisterAirtimeServiceFactory can be used by outside callers to register a airtime factory
+// for use by the engine
+func RegisterAirtimeServiceFactory(factory engine.AirtimeServiceFactory) {
+	airtimeFactory = factory
+}
+
+// Engine returns the global engine instance for use with real sessions
 func Engine() flows.Engine {
 	engInit.Do(func() {
+		webhookHeaders := map[string]string{
+			"User-Agent":      "RapidProMailroom/" + config.Mailroom.Version,
+			"X-Mailroom-Mode": "normal",
+		}
+
 		eng = engine.NewBuilder().
-			WithDefaultUserAgent("RapidProMailroom/" + config.Mailroom.Version).
+			WithWebhookServiceFactory(webhooks.NewServiceFactory(httpClient, webhookHeaders, 10000)).
+			WithClassificationServiceFactory(classificationFactory).
+			WithAirtimeServiceFactory(airtimeFactory).
 			WithMaxStepsPerSprint(config.Mailroom.MaxStepsPerSprint).
 			Build()
 	})
@@ -20,5 +65,46 @@ func Engine() flows.Engine {
 	return eng
 }
 
-var eng flows.Engine
-var engInit sync.Once
+// Simulator returns the global engine instance for use with simulated sessions
+func Simulator() flows.Engine {
+	simulatorInit.Do(func() {
+		webhookHeaders := map[string]string{
+			"User-Agent":      "RapidProMailroom/" + config.Mailroom.Version,
+			"X-Mailroom-Mode": "simulation",
+		}
+
+		simulator = engine.NewBuilder().
+			WithWebhookServiceFactory(webhooks.NewServiceFactory(httpClient, webhookHeaders, 10000)).
+			WithClassificationServiceFactory(classificationFactory).   // simulated sessions do real classification
+			WithAirtimeServiceFactory(simulatorAirtimeServiceFactory). // but faked airtime transfers
+			WithMaxStepsPerSprint(config.Mailroom.MaxStepsPerSprint).
+			Build()
+	})
+
+	return simulator
+}
+
+func simulatorAirtimeServiceFactory(session flows.Session) (flows.AirtimeService, error) {
+	return &simulatorAirtimeService{}, nil
+}
+
+type simulatorAirtimeService struct{}
+
+func (s *simulatorAirtimeService) Transfer(session flows.Session, sender urns.URN, recipient urns.URN, amounts map[string]decimal.Decimal, logHTTP flows.HTTPLogCallback) (*flows.AirtimeTransfer, error) {
+	transfer := &flows.AirtimeTransfer{
+		Sender:        sender,
+		Recipient:     recipient,
+		DesiredAmount: decimal.Zero,
+		ActualAmount:  decimal.Zero,
+	}
+
+	// pick arbitrary currency/amount pair in map
+	for currency, amount := range amounts {
+		transfer.Currency = currency
+		transfer.DesiredAmount = amount
+		transfer.ActualAmount = amount
+		break
+	}
+
+	return transfer, nil
+}