+++ ivr/twiml/twiml.go
@@ -9,6 +9,7 @@ import (
 	"encoding/json"
 	"encoding/xml"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"net/http"
 	"net/url"
@@ -154,8 +155,10 @@ func (c *client) RequestCall(client *http.Client, number urns.URN, callbackURL s
 	if err != nil {
 		return ivr.NilCallID, errors.Wrapf(err, "error trying to start call")
 	}
+	defer resp.Body.Close()
 
 	if resp.StatusCode != 201 {
+		io.Copy(ioutil.Discard, resp.Body)
 		return ivr.NilCallID, errors.Errorf("received non 201 status for call start: %d", resp.StatusCode)
 	}
 
@@ -191,6 +194,8 @@ func (c *client) HangupCall(client *http.Client, callID string) error {
 	if err != nil {
 		return errors.Wrapf(err, "error trying to hangup call")
 	}
+	defer resp.Body.Close()
+	io.Copy(ioutil.Discard, resp.Body)
 
 	if resp.StatusCode != 200 {
 		return errors.Errorf("received non 204 trying to hang up call: %d", resp.StatusCode)