+++ ivr/twiml/twiml_test.go
@@ -36,40 +36,40 @@ func TestResponseForSprint(t *testing.T) {
 		Expected string
 	}{
 		{
-			[]flows.Event{events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "hello world", nil, nil, nil))},
+			[]flows.Event{events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "hello world", nil, nil, nil))},
 			nil,
 			`<Response><Say>hello world</Say><Hangup></Hangup></Response>`,
 		},
 		{
-			[]flows.Event{events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "hello world", []utils.Attachment{utils.Attachment("audio:/recordings/foo.wav")}, nil, nil))},
+			[]flows.Event{events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "hello world", []utils.Attachment{utils.Attachment("audio:/recordings/foo.wav")}, nil, nil))},
 			nil,
 			`<Response><Play>https://mailroom.io/recordings/foo.wav</Play><Hangup></Hangup></Response>`,
 		},
 		{
-			[]flows.Event{events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "hello world", []utils.Attachment{utils.Attachment("audio:https://temba.io/recordings/foo.wav")}, nil, nil))},
+			[]flows.Event{events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "hello world", []utils.Attachment{utils.Attachment("audio:https://temba.io/recordings/foo.wav")}, nil, nil))},
 			nil,
 			`<Response><Play>https://temba.io/recordings/foo.wav</Play><Hangup></Hangup></Response>`,
 		},
 		{
 			[]flows.Event{
-				events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "hello world", nil, nil, nil)),
-				events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "goodbye", nil, nil, nil)),
+				events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "hello world", nil, nil, nil)),
+				events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "goodbye", nil, nil, nil)),
 			},
 			nil,
 			`<Response><Say>hello world</Say><Say>goodbye</Say><Hangup></Hangup></Response>`,
 		},
 		{
-			[]flows.Event{events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "enter a number", nil, nil, nil))},
+			[]flows.Event{events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "enter a number", nil, nil, nil))},
 			waits.NewActivatedMsgWait(nil, hints.NewFixedDigitsHint(1)),
 			`<Response><Gather numDigits="1" timeout="30" action="http://temba.io/resume?session=1&amp;wait_type=gather"><Say>enter a number</Say></Gather><Redirect>http://temba.io/resume?session=1&amp;wait_type=gather&amp;timeout=true</Redirect></Response>`,
 		},
 		{
-			[]flows.Event{events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "enter a number, then press #", nil, nil, nil))},
+			[]flows.Event{events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "enter a number, then press #", nil, nil, nil))},
 			waits.NewActivatedMsgWait(nil, hints.NewTerminatedDigitsHint("#")),
 			`<Response><Gather finishOnKey="#" timeout="30" action="http://temba.io/resume?session=1&amp;wait_type=gather"><Say>enter a number, then press #</Say></Gather><Redirect>http://temba.io/resume?session=1&amp;wait_type=gather&amp;timeout=true</Redirect></Response>`,
 		},
 		{
-			[]flows.Event{events.NewIVRCreatedEvent(flows.NewMsgOut(urn, channelRef, "say something", nil, nil, nil))},
+			[]flows.Event{events.NewIVRCreated(flows.NewMsgOut(urn, channelRef, "say something", nil, nil, nil))},
 			waits.NewActivatedMsgWait(nil, hints.NewAudioHint()),
 			`<Response><Say>say something</Say><Record action="http://temba.io/resume?session=1&amp;wait_type=record" maxLength="600"></Record><Redirect>http://temba.io/resume?session=1&amp;wait_type=record&amp;empty=true</Redirect></Response>`,
 		},