+++ models/assets.go
@@ -12,8 +12,10 @@ import (
 	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/flows/engine"
+	"github.com/nyaruka/mailroom/goflow"
 	cache "github.com/patrickmn/go-cache"
 	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 )
 
 // OrgAssets is our top level cache of all things contained in an org. It is used to build
@@ -36,6 +38,9 @@ type OrgAssets struct {
 	channelsByID   map[ChannelID]*Channel
 	channelsByUUID map[assets.ChannelUUID]*Channel
 
+	classifiers       []assets.Classifier
+	classifiersByUUID map[assets.ClassifierUUID]*Classifier
+
 	campaigns             []*Campaign
 	campaignEventsByField map[FieldID][]*CampaignEvent
 	campaignEventsByID    map[CampaignEventID]*CampaignEvent
@@ -55,6 +60,7 @@ type OrgAssets struct {
 	resthooks []assets.Resthook
 	templates []assets.Template
 	triggers  []*Trigger
+	globals   []assets.Global
 
 	locations        []assets.LocationHierarchy
 	locationsBuiltAt time.Time
@@ -86,6 +92,8 @@ func NewOrgAssets(ctx context.Context, db *sqlx.DB, orgID OrgID, prev *OrgAssets
 
 		channelsByID:   make(map[ChannelID]*Channel),
 		channelsByUUID: make(map[assets.ChannelUUID]*Channel),
+
+		classifiersByUUID: make(map[assets.ClassifierUUID]*Classifier),
 
 		fieldsByUUID: make(map[assets.FieldUUID]*Field),
 		fieldsByKey:  make(map[string]*Field),
@@ -119,6 +127,14 @@ func NewOrgAssets(ctx context.Context, db *sqlx.DB, orgID OrgID, prev *OrgAssets
 		channel := c.(*Channel)
 		o.channelsByID[channel.ID()] = channel
 		o.channelsByUUID[channel.UUID()] = channel
+	}
+
+	o.classifiers, err = loadClassifiers(ctx, db, orgID)
+	if err != nil {
+		return nil, errors.Wrapf(err, "error loading classifier assets for org %d", orgID)
+	}
+	for _, c := range o.classifiers {
+		o.classifiersByUUID[c.UUID()] = c.(*Classifier)
 	}
 
 	o.fields, err = loadFields(ctx, db, orgID)
@@ -176,6 +192,11 @@ func NewOrgAssets(ctx context.Context, db *sqlx.DB, orgID OrgID, prev *OrgAssets
 		return nil, errors.Wrapf(err, "error loading templates for org %d", orgID)
 	}
 
+	o.globals, err = loadGlobals(ctx, db, orgID)
+	if err != nil {
+		return nil, errors.Wrapf(err, "error loading globals for org %d", orgID)
+	}
+
 	// cache locations for an hour
 	if prev != nil && time.Since(prev.locationsBuiltAt) < locationCacheTimeout {
 		o.locations = prev.locations
@@ -225,11 +246,7 @@ func GetOrgAssets(ctx context.Context, db *sqlx.DB, orgID OrgID) (*OrgAssets, er
 
 // NewSessionAssets creates new sessions assets, returning the result
 func NewSessionAssets(org *OrgAssets) (flows.SessionAssets, error) {
-	assets, err := engine.NewSessionAssets(org)
-	if err != nil {
-		return nil, err
-	}
-	return assets, nil
+	return engine.NewSessionAssets(org, goflow.MigrationConfig())
 }
 
 // GetSessionAssets returns a goflow session assets object for the passed in org assets
@@ -271,6 +288,14 @@ func (a *OrgAssets) ChannelByID(channelID ChannelID) *Channel {
 func (a *OrgAssets) AddTestChannel(channel assets.Channel) {
 	a.channels = append(a.channels, channel)
 	// we don't populate our maps for uuid or id, shouldn't be used in any hook anyways
+}
+
+func (a *OrgAssets) Classifiers() ([]assets.Classifier, error) {
+	return a.classifiers, nil
+}
+
+func (a *OrgAssets) ClassifierByUUID(classifierUUID assets.ClassifierUUID) *Classifier {
+	return a.classifiersByUUID[classifierUUID]
 }
 
 func (a *OrgAssets) Fields() ([]assets.Field, error) {
@@ -345,6 +370,9 @@ func (a *OrgAssets) SetFlow(flowID FlowID, flow flows.Flow) (*Flow, error) {
 		return nil, errors.Wrapf(err, "error marshalling flow definition")
 	}
 
+	logrus.WithField("flow_id", flowID).WithField("flow_uuid", flow.UUID()).Debug("set debug flow")
+	fmt.Println(string(definition))
+
 	f := &Flow{}
 	f.f.UUID = flow.UUID()
 	f.f.Name = flow.Name()
@@ -416,4 +444,8 @@ func (a *OrgAssets) ResthookBySlug(slug string) *Resthook {
 
 func (a *OrgAssets) Templates() ([]assets.Template, error) {
 	return a.templates, nil
+}
+
+func (a *OrgAssets) Globals() ([]assets.Global, error) {
+	return a.globals, nil
 }