+++ CHANGELOG.md
@@ -1,3 +1,125 @@
+v5.3.6
+----------
+ * Load global assets from database
+
+v5.3.5
+----------
+ * Update to goflow v0.59.0
+
+v5.3.4
+----------
+ * Update to goflow v0.58.0
+
+v5.3.3
+----------
+ * Update to goflow v0.56.3
+
+v5.3.2
+----------
+ * Update to goflow v0.56.2
+
+v5.3.1
+----------
+ * Update to latest goflow
+
+v5.3.0
+----------
+ * Use github actions
+
+v5.2.4
+----------
+ * Add new dump file with bothub instead of bh
+
+v5.2.3
+----------
+ * collect queue size 15 seconds after the minute
+
+v5.2.2
+----------
+ * Update to goflow v0.55.0
+
+v5.2.1
+----------
+ * Fix NPE when IVR channel not found
+
+v5.2.0
+----------
+ * RapidPro 5.2 release
+
+v2.1.0
+----------
+ * Update to goflow v0.54.1
+
+v2.0.44
+----------
+ * fail calls that have missing channels when retrying them
+
+v2.0.43
+----------
+ * Simulation should use a different engine instance with a fake airtime service
+ * Add bothub classifier service factory
+
+v2.0.42
+----------
+ * Update to goflow v0.54.0
+
+v2.0.41
+----------
+ * only release on PG10
+
+v2.0.40
+----------
+ * Implement hook for airtime_transferred event
+ * Use DTOne for airtime service if configured
+
+v2.0.39
+----------
+ * Update to latest Librato library
+ * Audit closing HTTP bodies
+ * Add smtp-test command
+
+v2.0.38
+----------
+ * Latest GoFlow
+ * Add Classifier / NLU support for LUIS and Wit.ai
+
+v2.0.37
+----------
+ * Schedules fired in Mailroom
+
+v2.0.36
+----------
+ * Fire schedules from Mailroom
+
+v2.0.35
+----------
+ * Log query execution and elapsed time
+ * Update to goflow v0.50.4
+
+v2.0.34
+----------
+ * Update to latest goflow v0.50.2
+ * Add support for triggering sessions via query within a flow
+
+v2.0.33
+----------
+ * Stop writing to legacy engine fields on flows_flowrun
+ * Move tasks into their own package
+ * Add ElasticSearch URL to README
+
+v2.0.32
+----------
+ * Add expression/migrate endpoint
+
+v2.0.31
+----------
+ * Allow interrupting sessions by flow
+ * Update to goflow v0.49.0
+
+v2.0.30
+----------
+ * Update to goflow v0.47.3
+
 v2.0.29
 ----------
  * Expire runs and sessions in a transaction to guarantee they are always in sync